+++ packages/core/src/state-management/repository.ts
@@ -3,7 +3,7 @@ import {
   AnchorEvent,
   AnchorOpts,
   AnchorStatus,
-  CommitType,
+  EventType,
   CreateOpts,
   DiagnosticsLogger,
   LoadOpts,
@@ -33,7 +33,7 @@ import { StreamUpdater } from '../stream-loading/stream-updater.js'
 import { CID } from 'multiformats/cid'
 import type { AnchorLoopHandler, AnchorService } from '../anchor/anchor-service.js'
 import type { AnchorRequestCarBuilder } from '../anchor/anchor-request-car-builder.js'
-import { AnchorRequestStatusName } from '@ceramicnetwork/codecs'
+import { AnchorRequestStatusName } from '@ceramicnetwork/common'
 import { CAR } from 'cartonne'
 import { FeedDocument, type Feed } from '../feed.js'
 import { doNotWait } from '../ancillary/do-not-wait.js'
@@ -72,7 +72,7 @@ function shouldIndex(state$: RunningState, index: LocalIndexApi): boolean {
 function commitAtTime(state: StreamState, timestamp: number): CommitID {
   let commitCid: CID = state.log[0].cid
   for (const entry of state.log) {
-    if (entry.type === CommitType.ANCHOR) {
+    if (entry.type === EventType.TIME) {
       if (entry.timestamp <= timestamp) {
         commitCid = entry.cid
       } else {
@@ -961,7 +961,7 @@ export class Repository {
     // TODO(NET-1614) Test that the timestamps are correctly passed to the Index API.
     const lastAnchor = asDate(StreamUtils.anchorTimestampFromState(state$.value))
     const firstAnchor = asDate(
-      state$.value.log.find((log) => log.type == CommitType.ANCHOR)?.timestamp
+      state$.value.log.find((log) => log.type == EventType.TIME)?.timestamp
     )
     const streamContent = {
       model: state$.value.metadata.model,
@@ -971,7 +971,6 @@ export class Repository {
       tip: state$.tip,
       lastAnchor: lastAnchor,
       firstAnchor: firstAnchor,
-      shouldIndex: state$.value.metadata.shouldIndex,
     }
 
     await this.index.indexStream(streamContent)