+++ packages/stream-caip10-link-handler/src/caip10-link-handler.ts
@@ -3,7 +3,7 @@ import { Caip10Link } from '@ceramicnetwork/stream-caip10-link'
 import {
   AnchorStatus,
   CommitData,
-  CommitType,
+  EventType,
   SignatureStatus,
   StreamConstructor,
   StreamHandler,
@@ -75,7 +75,7 @@ export class Caip10LinkHandler implements StreamHandler<Caip10Link> {
       metadata,
       signature: SignatureStatus.GENESIS,
       anchorStatus: AnchorStatus.NOT_REQUESTED,
-      log: [{ cid: commitData.cid, type: CommitType.GENESIS }],
+      log: [{ cid: commitData.cid, type: EventType.INIT }],
     }
 
     return state
@@ -129,7 +129,7 @@ export class Caip10LinkHandler implements StreamHandler<Caip10Link> {
         `Address '${legacyAccountCaip10.toLowerCase()}' used to sign update to Caip10Link doesn't match stream controller '${legacyControllerCaip10.toLowerCase()}'`
       )
     }
-    state.log.push({ cid: commitData.cid, type: CommitType.SIGNED })
+    state.log.push({ cid: commitData.cid, type: EventType.DATA })
     return {
       ...state,
       signature: SignatureStatus.SIGNED,