+++ packages/stream-model-instance-handler/src/__tests__/model-instance-document-handler.test.ts
@@ -7,7 +7,7 @@ import jsonpatch from 'fast-json-patch'
 import { ModelInstanceDocument } from '@ceramicnetwork/stream-model-instance'
 import type { ModelDefinition } from '@ceramicnetwork/stream-model'
 import {
-  CommitType,
+  EventType,
   StreamUtils,
   SignedCommitContainer,
   IpfsApi,
@@ -583,7 +583,7 @@ describe('ModelInstanceDocumentHandler', () => {
 
     const commitData = {
       cid: FAKE_CID_1,
-      type: CommitType.GENESIS,
+      type: EventType.INIT,
       commit: payload,
       envelope: commit.jws,
     }
@@ -605,7 +605,7 @@ describe('ModelInstanceDocumentHandler', () => {
 
     const commitData = {
       cid: FAKE_CID_BLOB,
-      type: CommitType.GENESIS,
+      type: EventType.INIT,
       commit: payload,
       envelope: commit.jws,
     }
@@ -624,7 +624,7 @@ describe('ModelInstanceDocumentHandler', () => {
 
     const commitData = {
       cid: FAKE_CID_1,
-      type: CommitType.GENESIS,
+      type: EventType.INIT,
       commit,
     }
     await expect(handler.applyCommit(commitData, context)).rejects.toThrow(
@@ -642,7 +642,7 @@ describe('ModelInstanceDocumentHandler', () => {
 
     const commitData = {
       cid: FAKE_CID_1,
-      type: CommitType.GENESIS,
+      type: EventType.INIT,
       commit,
     }
     expect(commitData.commit.data).toBeNull()
@@ -665,7 +665,7 @@ describe('ModelInstanceDocumentHandler', () => {
 
     const commitData = {
       cid: FAKE_CID_1,
-      type: CommitType.GENESIS,
+      type: EventType.INIT,
       commit: payload,
       envelope: commit.jws,
     }
@@ -687,7 +687,7 @@ describe('ModelInstanceDocumentHandler', () => {
 
     const commitData = {
       cid: FAKE_CID_1,
-      type: CommitType.GENESIS,
+      type: EventType.INIT,
       commit: payload,
       envelope: commit.jws,
     }
@@ -709,7 +709,7 @@ describe('ModelInstanceDocumentHandler', () => {
 
     const commitData = {
       cid: FAKE_CID_1,
-      type: CommitType.GENESIS,
+      type: EventType.INIT,
       commit: payload,
       envelope: commit.jws,
     }
@@ -731,7 +731,7 @@ describe('ModelInstanceDocumentHandler', () => {
 
     const commitData = {
       cid: FAKE_CID_1,
-      type: CommitType.GENESIS,
+      type: EventType.INIT,
       commit: payload,
       envelope: commit.jws,
     }
@@ -756,7 +756,7 @@ describe('ModelInstanceDocumentHandler', () => {
 
     const commitData = {
       cid: FAKE_CID_1,
-      type: CommitType.GENESIS,
+      type: EventType.INIT,
       commit: payload,
       envelope: commit.jws,
     }
@@ -778,7 +778,7 @@ describe('ModelInstanceDocumentHandler', () => {
 
     const genesisCommitData = {
       cid: FAKE_CID_1,
-      type: CommitType.GENESIS,
+      type: EventType.INIT,
       commit: payload,
       envelope: genesisCommit.jws,
     }
@@ -815,7 +815,7 @@ describe('ModelInstanceDocumentHandler', () => {
     // apply genesis
     const genesisCommitData = {
       cid: FAKE_CID_1,
-      type: CommitType.GENESIS,
+      type: EventType.INIT,
       commit: payload,
       envelope: genesisCommit.jws,
     }
@@ -838,7 +838,7 @@ describe('ModelInstanceDocumentHandler', () => {
     // apply signed
     const signedCommitData = {
       cid: FAKE_CID_2,
-      type: CommitType.SIGNED,
+      type: EventType.DATA,
       commit: sPayload,
       envelope: signedCommit.jws,
     }
@@ -859,7 +859,7 @@ describe('ModelInstanceDocumentHandler', () => {
     // apply genesis
     const genesisCommitData = {
       cid: FAKE_CID_1,
-      type: CommitType.GENESIS,
+      type: EventType.INIT,
       commit: genesisCommit,
     }
     let state = await handler.applyCommit(genesisCommitData, context)
@@ -880,7 +880,7 @@ describe('ModelInstanceDocumentHandler', () => {
 
     const signedCommitDataFail = {
       cid: FAKE_CID_2,
-      type: CommitType.SIGNED,
+      type: EventType.DATA,
       commit: payloadFail,
       envelope: signedCommitFail.jws,
     }
@@ -901,7 +901,7 @@ describe('ModelInstanceDocumentHandler', () => {
 
     const signedCommitDataOK = {
       cid: FAKE_CID_3,
-      type: CommitType.SIGNED,
+      type: EventType.DATA,
       commit: payloadOK,
       envelope: signedCommitOK.jws,
     }
@@ -921,7 +921,7 @@ describe('ModelInstanceDocumentHandler', () => {
     // apply genesis
     const genesisCommitData = {
       cid: FAKE_CID_1,
-      type: CommitType.GENESIS,
+      type: EventType.INIT,
       commit: genesisCommit,
     }
 
@@ -943,7 +943,7 @@ describe('ModelInstanceDocumentHandler', () => {
 
     const signedCommitData = {
       cid: FAKE_CID_2,
-      type: CommitType.SIGNED,
+      type: EventType.DATA,
       commit: payload,
       envelope: signedCommit.jws,
     }
@@ -967,7 +967,7 @@ describe('ModelInstanceDocumentHandler', () => {
     // apply genesis
     const genesisCommitData = {
       cid: FAKE_CID_1,
-      type: CommitType.GENESIS,
+      type: EventType.INIT,
       commit: payload,
       envelope: genesisCommit.jws,
     }
@@ -989,7 +989,7 @@ describe('ModelInstanceDocumentHandler', () => {
     // apply signed
     const signedCommitData_1 = {
       cid: FAKE_CID_2,
-      type: CommitType.SIGNED,
+      type: EventType.DATA,
       commit: sPayload1,
       envelope: signedCommit1.jws,
     }
@@ -1012,7 +1012,7 @@ describe('ModelInstanceDocumentHandler', () => {
     // apply signed
     const signedCommitData_2 = {
       cid: FAKE_CID_3,
-      type: CommitType.SIGNED,
+      type: EventType.DATA,
       commit: sPayload2,
       envelope: signedCommit2.jws,
     }
@@ -1163,7 +1163,7 @@ describe('ModelInstanceDocumentHandler', () => {
 
     const commitData = {
       cid: FAKE_CID_1,
-      type: CommitType.GENESIS,
+      type: EventType.INIT,
       commit: payload,
       envelope: commit.jws,
     }
@@ -1194,7 +1194,7 @@ describe('ModelInstanceDocumentHandler', () => {
     // apply genesis
     const genesisCommitData = {
       cid: FAKE_CID_1,
-      type: CommitType.GENESIS,
+      type: EventType.INIT,
       commit: payload,
       envelope: genesisCommit.jws,
     }
@@ -1217,7 +1217,7 @@ describe('ModelInstanceDocumentHandler', () => {
     // apply signed
     const signedCommitData = {
       cid: FAKE_CID_2,
-      type: CommitType.SIGNED,
+      type: EventType.DATA,
       commit: sPayload,
       envelope: signedCommit.jws,
     }
@@ -1239,7 +1239,7 @@ describe('ModelInstanceDocumentHandler', () => {
 
     const genesisCommitData = {
       cid: FAKE_CID_1,
-      type: CommitType.GENESIS,
+      type: EventType.INIT,
       commit: payload,
       envelope: genesisCommit.jws,
       timestamp: Date.now(),
@@ -1264,7 +1264,7 @@ describe('ModelInstanceDocumentHandler', () => {
     // apply genesis
     const genesisCommitData = {
       cid: FAKE_CID_1,
-      type: CommitType.GENESIS,
+      type: EventType.INIT,
       commit: payload,
       envelope: genesisCommit.jws,
     }
@@ -1285,12 +1285,12 @@ describe('ModelInstanceDocumentHandler', () => {
     // apply signed
     const signedCommitData = {
       cid: FAKE_CID_2,
-      type: CommitType.SIGNED,
+      type: EventType.DATA,
       commit: sPayload,
       envelope: signedCommit.jws,
     }
     await expect(handler.applyCommit(signedCommitData, context, state)).rejects.toThrow(
-      `Unsupported metadata changes for ModelInstanceDocument Stream ${doc.id}: controllers. Only the shouldIndex argument can be changed.`
+      /Updating metadata for ModelInstanceDocument Streams is not allowed/
     )
   })
 
@@ -1308,7 +1308,7 @@ describe('ModelInstanceDocumentHandler', () => {
     // apply genesis
     const genesisCommitData = {
       cid: FAKE_CID_1,
-      type: CommitType.GENESIS,
+      type: EventType.INIT,
       commit: payload,
       envelope: genesisCommit.jws,
     }
@@ -1328,7 +1328,7 @@ describe('ModelInstanceDocumentHandler', () => {
     // apply signed
     const signedCommitData = {
       cid: FAKE_CID_2,
-      type: CommitType.SIGNED,
+      type: EventType.DATA,
       commit: sPayload,
       envelope: signedCommit.jws,
     }
@@ -1351,7 +1351,7 @@ describe('ModelInstanceDocumentHandler', () => {
     // apply genesis
     const genesisCommitData = {
       cid: FAKE_CID_1,
-      type: CommitType.GENESIS,
+      type: EventType.INIT,
       commit: payload,
       envelope: genesisCommit.jws,
     }
@@ -1371,7 +1371,7 @@ describe('ModelInstanceDocumentHandler', () => {
     // apply signed
     const signedCommitData = {
       cid: FAKE_CID_2,
-      type: CommitType.SIGNED,
+      type: EventType.DATA,
       commit: sPayload,
       envelope: signedCommit.jws,
     }
@@ -1394,7 +1394,7 @@ describe('ModelInstanceDocumentHandler', () => {
     // apply genesis
     const genesisCommitData = {
       cid: FAKE_CID_1,
-      type: CommitType.GENESIS,
+      type: EventType.INIT,
       commit: payload,
       envelope: genesisCommit.jws,
     }
@@ -1417,7 +1417,7 @@ describe('ModelInstanceDocumentHandler', () => {
     // apply signed
     const signedCommitData = {
       cid: FAKE_CID_2,
-      type: CommitType.SIGNED,
+      type: EventType.DATA,
       commit: sPayload,
       envelope: signedCommit.jws,
     }
@@ -1430,7 +1430,7 @@ describe('ModelInstanceDocumentHandler', () => {
     await ipfs.dag.put(anchorProof, FAKE_CID_3)
     const anchorCommitData = {
       cid: FAKE_CID_4,
-      type: CommitType.ANCHOR,
+      type: EventType.TIME,
       commit: { proof: FAKE_CID_3, id: FAKE_CID_1, prev: FAKE_CID_2 },
       proof: anchorProof,
       timestamp: 1615799679,
@@ -1455,7 +1455,7 @@ describe('ModelInstanceDocumentHandler', () => {
     // genesis commit applied one hour before rotation
     const genesisCommitData = {
       cid: FAKE_CID_1,
-      type: CommitType.GENESIS,
+      type: EventType.INIT,
       commit: payload,
       envelope: genesisCommit.jws,
       timestamp: rotateDate.valueOf() / 1000 - 60 * 60,
@@ -1480,7 +1480,7 @@ describe('ModelInstanceDocumentHandler', () => {
 
     const signedCommitData = {
       cid: FAKE_CID_2,
-      type: CommitType.SIGNED,
+      type: EventType.DATA,
       commit: sPayload,
       envelope: signedCommit.jws,
       // 24 hours after rotation
@@ -1509,7 +1509,7 @@ describe('ModelInstanceDocumentHandler', () => {
     // commit is applied 1 hour before rotation
     const genesisCommitData = {
       cid: FAKE_CID_1,
-      type: CommitType.GENESIS,
+      type: EventType.INIT,
       commit: payload,
       envelope: genesisCommit.jws,
       timestamp: rotateDate.valueOf() / 1000 - 60 * 60,
@@ -1537,7 +1537,7 @@ describe('ModelInstanceDocumentHandler', () => {
     // commit is applied 1 hour after the rotation
     const genesisCommitData = {
       cid: FAKE_CID_1,
-      type: CommitType.GENESIS,
+      type: EventType.INIT,
       commit: payload,
       envelope: genesisCommit.jws,
       timestamp: Math.floor(rotateDate.valueOf() / 1000) + 60 * 60,
@@ -1562,7 +1562,7 @@ describe('ModelInstanceDocumentHandler', () => {
 
     const commitData = {
       cid: FAKE_CID_1,
-      type: CommitType.GENESIS,
+      type: EventType.INIT,
       commit: payload,
       envelope: commit.jws,
     }
@@ -1585,7 +1585,7 @@ describe('ModelInstanceDocumentHandler', () => {
     // apply genesis
     const genesisCommitData = {
       cid: FAKE_CID_1,
-      type: CommitType.GENESIS,
+      type: EventType.INIT,
       commit: payload,
       envelope: genesisCommit.jws,
     }
@@ -1608,7 +1608,7 @@ describe('ModelInstanceDocumentHandler', () => {
     // apply genesis
     const genesisCommitData = {
       cid: FAKE_CID_1,
-      type: CommitType.GENESIS,
+      type: EventType.INIT,
       commit: payload,
       envelope: genesisCommit.jws,
     }
@@ -1629,7 +1629,7 @@ describe('ModelInstanceDocumentHandler', () => {
     // apply genesis
     const genesisCommitData = {
       cid: FAKE_CID_1,
-      type: CommitType.GENESIS,
+      type: EventType.INIT,
       commit: payload,
       envelope: genesisCommit.jws,
     }
@@ -1650,7 +1650,7 @@ describe('ModelInstanceDocumentHandler', () => {
     // apply genesis
     const genesisCommitData = {
       cid: FAKE_CID_1,
-      type: CommitType.GENESIS,
+      type: EventType.INIT,
       commit: payload,
       envelope: genesisCommit.jws,
     }
@@ -1671,7 +1671,7 @@ describe('ModelInstanceDocumentHandler', () => {
     // apply genesis
     const genesisCommitData = {
       cid: FAKE_CID_1,
-      type: CommitType.GENESIS,
+      type: EventType.INIT,
       commit: payload,
       envelope: genesisCommit.jws,
     }